Added the INSTALL TARGET to the CMakeLists of the GEOtop libraries:
    1. gt_utilities
    2. libraries/ascii
    3. libraries/fluidturtle
    4. meteoio_plugin
    5. netCDF

The make install command is still not working, but this modify allows to install all the GEOtop libraries inside the folder /DEFINED/PATH/geotop/. Although, these libraries have to be linked to the executable so that it could find them. To do this, setting the export LD_LIBRARY_PATH=/DEFINED/PATH/geotop:${LD_LIBRARY_PATH} adding it to your .bash_profile is required.

ATTENTION!!! GEOtop IS STILL NOT WORKING!